requirements.txt
flask
pillow
pymupdf
pytesseract
opencv-python
numpy
requests
